All shared structs, consts and hard-coded variables now in lib.rs
Moved hard-coded path to lib.rs

Hard-coded path is in lib.rs, also moved final consts from main.rs to lib.rs
NOTE - one hard-coded path still exists in external/enarx-keep@.service,
probably needs build.rs

Moved hard-coded path from keep-loader.rs to lib.rs
