+++ host-components/src/main.rs
@@ -1,6 +1,7 @@
 #[macro_use]
 extern crate serde_derive;
 
+use ::host_components::*;
 use rand::Rng;
 use std::collections::HashMap;
 use std::fs::File;
@@ -9,10 +10,6 @@ use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use std::path::Path;
 use warp::http::StatusCode;
 use warp::Filter;
-
-pub const PROTO_VERSION: f32 = 0.1;
-pub const PROTO_NAME: &str = "Enarx-Keep-Manager";
-pub const BIND_PORT: u16 = 3030;
 
 #[tokio::main]
 async fn main() {
@@ -51,8 +48,8 @@ async fn main() {
         .await;
 }
 
-//TODO: move this to lib.rs or similar
 mod models {
+    use ::host_components::*;
     use glob::glob;
     use serde_derive::{Deserialize, Serialize};
     use std::fs;
@@ -61,39 +58,6 @@ mod models {
     use std::path::Path;
     use std::sync::Arc;
     use tokio::sync::Mutex;
-
-    //TODO - put this in the right place
-    pub const KEEP_INFO_COMMAND: &str = "keep-info";
-
-    pub type KeepLoaderList = Arc<Mutex<Vec<KeepLoader>>>;
-
-    #[derive(Serialize, Deserialize, Clone)]
-    pub struct KeepLoader {
-        pub state: u8,
-        pub kuuid: usize,
-        pub app_loader_bind_port: u16,
-        pub bindaddress: String,
-        //we may wish to add information here about whether we're happy to share
-        // all of this information with external parties, but since the keeploader
-        // is operating outside the TEE boundary, there's only so much we can do
-        // to keep this information confidential
-    }
-
-    #[derive(Serialize, Deserialize)]
-    pub struct JsonCommand {
-        pub commandtype: String,
-        pub commandcontents: String,
-    }
-
-    #[derive(Serialize, Deserialize, Clone)]
-    pub struct KeepLoaderVec {
-        pub klvec: Vec<KeepLoader>,
-    }
-
-    #[derive(Serialize, Deserialize, Clone)]
-    pub struct UndefinedReply {
-        pub text: String,
-    }
 
     pub fn new_empty_KeepLoaderList() -> KeepLoaderList {
         Arc::new(Mutex::new(Vec::new()))
@@ -164,7 +128,8 @@ mod models {
 }
 
 mod filters {
-    use super::models::{JsonCommand, KeepLoader, KeepLoaderList, KeepLoaderVec, UndefinedReply};
+    use ::host_components::*;
+    //    use super::models::{JsonCommand, KeepLoader, KeepLoaderList, KeepLoaderVec, UndefinedReply};
     use fork::{daemon, setsid, Fork};
     use rand::Rng;
     use serde_json::json;
@@ -176,22 +141,6 @@ mod filters {
     use std::process::{Command, Stdio};
     //    use subprocess::Exec;
     use warp::Filter;
-
-    //there's got to be a better way of doing this (enums?) in a lib.rs file?
-    pub const CONTRACT_COMMAND: &str = "command";
-    pub const KEEP_COMMAND: &str = "command";
-    pub const KEEP_AUTH: &str = "auth-token";
-    pub const KEEP_KUUID: &str = "kuuid";
-    pub const KEEP_ARCH: &str = "keep-arch";
-    pub const KEEP_ARCH_WASI: &str = "wasi";
-    pub const KEEP_ARCH_SEV: &str = "AMD-SEV";
-    pub const KEEP_ARCH_SGX: &str = "Intel-SGX";
-    pub const KEEP_APP_LOADER_BIND_PORT: &str = "app-loader-bind-port";
-    pub const APP_LOADER_BIND_PORT_START: u16 = 3031;
-    pub const KEEP_APP_LOADER_STATE_UNDEF: u8 = 0;
-    pub const KEEP_INFO_COMMAND: &str = "keep-info";
-
-    pub const KEEP_APP_LOADER_START_COMMAND: &str = "apploader-start"; // requires app_loader_bind_port to be provided
 
     pub fn with_keeploaderlist(
         keeploaderlist: KeepLoaderList,
